pyproject.toml
[project]
name = "nonebot-plugin-kawaii-status"
version = "0.1.0"
description = "NoneBot2 服务器状态查看插件 / View server status for NoneBot2"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "KomoriDev", email = "mute231010@gmail.com" },
]
dependencies = [
    "nonebot-plugin-alconna>=0.37.0",
    "psutil>=5.9.8",
    "pillow>=10.2.0",
    "nonebot2[fastapi]>=2.2.0",
]
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/KomoriDev/nonebot-plugin-kawaii-status"
repository = "https://github.com/KomoriDev/nonebot-plugin-kawaii-status"

[project.optional-dependencies]
adapters = [
    "nonebot-adapter-onebot>=2.4.1",
]

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.2.1",
    "isort>=5.13.2",
    "black>=24.2.0",
]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]


[tool.ruff]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]
line-length = 120
target-version = "py38"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
